requirements.txt
fastapi
uvicorn
streamlit
python-multipart
langchain-google-genai
langchain
python-dotenv
pypdf
requests